### 2023-05-04 #5438 @kthyng
Hi @SimonGoring and thanks for your submission! I see there is a link for docs, but it points to the same info as in the readme. How does a person see the docs or vignettes from your package? I am not an R person.
