### 2023-05-08 #5438 @xuanxu
The error compiling the PDF is caused by the identation in the metadata of the paper.md file.
Merging [this change](https://github.com/NeotomaDB/neotoma2/pull/24) should fix it.
