### 2023-05-10 #5438 @SimonGoring
@kthyng The vignettes come as part of the R package, but can also be seen here (as a markdown document):
https://github.com/NeotomaDB/neotoma2/blob/production/vignettes/neotoma2-package.Rmd

We also have documentation here:
https://open.neotomadb.org/Current_Workshop/simple_workflow.html

And the compiled PDF is here:
https://cran.r-project.org/web/packages/neotoma2/neotoma2.pdf
