### 2023-05-19 a6dfd14
Cleaning up the repo with new pubs and some markdown changes.
Added a set of publications to paper.bib and added DOIs where neccessary. Fixed those references in the paper.md file.
Added affiliations for researchers in the md file.
Removed LICENSE to keep the LICENSE.md file.
