+++ .github/workflows/draft-pdf.yml
@@ -1,5 +1,4 @@
 on: [push]
-
 jobs:
   paper:
     runs-on: ubuntu-latest
