### 2023-05-23 c74e263
added try statements. updated description and cran comments.
