+++ NEWS.md
@@ -1,3 +1,8 @@
-# neotoma2 1.0.0
+# neotoma2 1.0.1
 
 * Added a `NEWS.md` file to track changes to the package.
+
+---
+
+May 23rd, 2023
+Added handling of SSL certificate errors.
