### 2023-06-13 d7fece4
Fixed bug that was only parsing a set of the results
