### 2023-06-14 #5438 @mhesselbarth
I don't have a background in Paleoecology, but I could give it a try to review 👍
