### 2023-06-14 #5561 @editorialbot
[REVIEW]: neotoma2: An R package to access data from the Neotoma Paleoecology Database
**Submitting author:** @SimonGoring (<a href="http://orcid.org/0000-0002-2700-4605">Simon Goring</a>)
**Repository:** https://github.com/NeotomaDB/neotoma2
**Branch with paper.md** (empty if default branch): production
**Version:** v1.0.3.1
