+++ NEWS.md
@@ -6,3 +6,6 @@
 
 May 23rd, 2023
 Added handling of SSL certificate errors.
+
+--- June 13th, 2023
+Added bug fixes that were trimming down the results.
