+++ R/get_downloads.R
@@ -151,13 +151,10 @@ parse_download <- function(result, verbose = TRUE) {
 #' @export
 get_downloads.numeric <- function(x, verbose = TRUE, ...) {
   
-  
   if (length(x) > 0) {
     dataset <- paste0(x, collapse = ",")
   }
-  
-  
-  
+
   base_url <- paste0("data/downloads/", dataset)
   result <- parseURL(base_url, ...) # nolint
   
@@ -206,12 +203,23 @@ get_downloads.sites <- function(x, verbose = TRUE, ...) {
   if('all_data' %in% names(cl)){
     all_data = cl$all_data
   }else{
-    all_data = TRUE
+    cl[['all_data']] = TRUE
+  }
+  
+  if('limit' %in% names(cl)){
+    cl[['all_data']] = FALSE
+  }
+  
+  if('offset' %in% names(cl)){
+    cl[['all_data']] = FALSE
   }
   ## Fixing all data line
   
-  output <- get_downloads(x = output, verbose, all_data = all_data)
+  cl[['x']] <- output
+  cl[['verbose']] <- verbose
   
+  output <- do.call(get_downloads, cl)
+
   return(output)
 }
 
@@ -231,7 +239,6 @@ get_downloads.character <- function(x, verbose = TRUE, ...) {
   result <- jsonlite::fromJSON(x,
                                flatten = FALSE,
                                simplifyVector = FALSE)
-  
   result <- result %>%
     cleanNULL()
   
