+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: neotoma2
 Title: Working with the Neotoma Paleoecology Database
 Date: 2023-06-13
-Version: 1.0.1
+Version: 1.0.2
 Authors@R:
     c(person(given = "Dominguez Vidana",
              family = "Socorro",
