### 2023-06-19 2de7d7a
modified parseURL so that it can take ids and a limit argument
