+++ R/get_sites.R
@@ -244,11 +244,9 @@ get_sites.default <- function(...) { # nolint
 #' }
 #' @export
 get_sites.numeric <- function(x, ...) {
-
   if (length(x) > 0) {
     siteids <- paste0(x, collapse = ",")
   }
-
   base_url <- paste0("data/sites/", siteids)
   result <- neotoma2::parseURL(base_url, ...)
   result_length <- length(result[2]$data)
@@ -293,6 +291,7 @@ get_sites.sites <- function(x, ...) {
       unique() %>%
       as.numeric() %>%
       na.omit() %>%
+      suppressWarnings() %>%
       paste0(., collapse = ",")
   }
  
@@ -317,7 +316,8 @@ get_sites.sites <- function(x, ...) {
   }
   ## Fixing all data line
   
-  cl[['x']] <- siteids
+  cl[['siteid']] <- siteids
+  cl[['x']] <- NULL
 
   output <- do.call(get_sites, cl)
   
