### 2023-06-20 #5561 @crvernon
@xuanxu will you assist with the above compile error?

It looks like compile fails here:

```
1.5.4/lib/theoj/submission.rb:57:in `article_metadata'
	from /home/runner/work/_actions/xuanxu/paper-action/main/get_paper.rb:24:in `<main>'
Error: Process completed with exit code 1.
```

You had fixed a similar issue [in this PR](https://github.com/NeotomaDB/neotoma2/pull/24) during the pre-review, but this doesn't seem to be the same thing.  And I don't believe the file has been altered since that PR was merged.

Target branch is set to `production`
