### 2023-06-20 #5561 @xuanxu
The error message is [`Problem with affiliations for Simon Goring, perhaps the affiliations index need quoting?`](https://github.com/openjournals/joss-papers/actions/runs/5322527417/jobs/9639096590#step:3:94)

I've checked and the format used in the paper is invalid, I've opened a PR that should fix it: https://github.com/NeotomaDB/neotoma2/pull/27
