+++ NEWS.md
@@ -7,6 +7,6 @@
 May 23rd, 2023
 Added handling of SSL certificate errors.
 
---- June 13th, 2023
+--- June 26th, 2023
 Added bug fixes that were trimming down the results.
 Parsed URLs different to handle get_downloads.sites or get_datasets.sites with limit parameters.
