### 2023-07-30 NeotomaDB/neotoma2#29 @kanishkan91
JOSS review: Build vignettes when installing package
When installing the package, the vignettes dont build automatically. This should be a relatively simple fix. All that is required is changing the README from-

devtools::install_github('NeotomaDB/neotoma2')

to

devtools::install_github('NeotomaDB/neotoma2', build_vignettes=TRUE)

UPDATE: I confirmed that the build_vignettes passes when indtalling. So, this can easily be updated in the README.

https://github.com/openjournals/joss-reviews/issues/5561
