### 2023-07-30 NeotomaDB/neotoma2#30 @kanishkan91
JOSS review: documentation does not include vignettes. These should be produced using pkgdown
The current docs- http://open.neotomadb.org/neotoma2/ dont include the vignettes. I think the authors would just have to do

`pkgdown::build_articles()` to generate these and include them on the GitHub docs.

https://github.com/openjournals/joss-reviews/issues/5561
