### 2023-07-30 NeotomaDB/neotoma2#31 @kanishkan91
JOSS review: Contribution instructions (add email)
The authors should add  an email which new users can reach out to in case of any questions/confusion

https://github.com/openjournals/joss-reviews/issues/5561
