### 2023-07-30 NeotomaDB/neotoma2#32 @kanishkan91
JOSS review: Could authors add state of the field to the paper?
Can the authors add some text regarding the current available options (other than your package) which can be used to explore the neotoma database? And why would this be much more accessible.

I'm guessing this is because using R is very user friendly and users can conviniently add their own sites, methods? Either way, this should be fairly simple to add to the paper itself.

https://github.com/openjournals/joss-reviews/issues/5561
