### 2023-07-30 NeotomaDB/neotoma2#33 @kanishkan91
JOSS review: (OPTIONAL DEVELOPMENT!) Add parallel processing to speed up get_sites()?
The get_sites() function is used to derive site specific info. It seems to call an API function. It can take very long to get multiple sites at once especially if the user provides an age filter.

Authors can consider just adding a parallelization here since this is just a fetch function. This could be as simple as below,

`  library(parallel)
  library(data.table)
#create cores and save to a cluster object

list_of_sites <- c(A,B,C)

parLapply(cluster,
          list_of_sites)->processed_sites

 processed_sites <- rbindlist(processed sites)

 `

Above is a crude example, but can work well. Again just an optional suggestion

https://github.com/openjournals/joss-reviews/issues/5561
