### 2023-07-30 #5561 @kanishkan91
@crvernon @SimonGoring - I have almost completed my review of this software package.

**Summary:** neotoma2 is a software package that can be used to query the neotoma database (which a database with several site specific paleontological and ecological data). This is clearly a very useful software package. In particular I commend the feature to query data using spatial information. My comments are rather minor and I will complete my checklist once the authors have addressed my comments. I opened issues for individual comments, but I will summarize the same below.

**Summary of issues**-

1. It seems the installation instructions right now don't install vignettes automatically. I would recommend that this be added to the installation instructions. See issue here- https://github.com/NeotomaDB/neotoma2/issues/29
2. Since there are vignettes available , the documentation on Github docs should include the same. This should be fairly simple to add. See issue- https://github.com/NeotomaDB/neotoma2/issues/30
3. Can the authors describe the state of the field in the paper? Just a paragraph describing why this package is better than what is already out there to query the neotoma database? I added some suggestions to the issue documented here- https://github.com/NeotomaDB/neotoma2/issues/32
4. Given this is a live package where users can add sites etc, can the authors report the total available sites that can be queried by this package? Ideally there should be a function that just reports this. Authors can also add this to a table in the README.
5. OPTIONAL FEATURE- It seems the "get" functions can get very slow when user queries several sites at once. Since the get functions are just calls via APIs, it should be fairly simple to add parallel processing so that this package can be much much faster. I have documented this here- https://github.com/NeotomaDB/neotoma2/issues/33
6. Can the authors add an email to the contribution instructions- https://github.com/NeotomaDB/neotoma2/issues/31

**Additional question**- While going through the vignette, I read that there used to be an older neotoma package. This new package is an improvement upon that. Can the authors explain what that means? Is that old package still available on GitHub? If this package is the latest and greatest, can the old package be deprecated? Perhaps all this is already answered somewhere and I just missed it, but just wanted to raise here.
