### 2023-07-30 NeotomaDB/neotoma2#34 @kanishkan91
JOSS review: Add a summary of current available sites that can be queried
Given this is a live package where users can add sites etc, can the authors report the total available sites that can be queried by this package? Ideally there should be a function that just reports this. Authors can also add this to a table in the README.

https://github.com/openjournals/joss-reviews/issues/5561
