### 2023-08-01 #5561 @sedv8808
Hello,

I believe that would be me, my GitHub profile is:
https://github.com/sedv8808

I have a compound last name so that makes it sometimes hard to find.

Thank you,

On Tue, Aug 1, 2023 at 8:27 AM Kanishka Narayan ***@***.***>
wrote:

> One other question. @SimonGoring <https://github.com/SimonGoring> The
> lead author on the paper, I could not find on GitHub. Could you comment on
> that? I would expect that the lead author of the paper would be a key
> contributor to the repo itself. Maybe I missed something (If so, my
> apologies!).
>
