+++ paper.md
@@ -35,7 +35,7 @@ The `neotoma2` R package has been under dynamic development for over a year, but
 
 The `neotoma` R package [@goring2015neotoma] leveraged the Neotoma Paleoeocology Database v1.0 API and had been one of the primary tools for researchers working with data from Neotoma [@wang2023plants;@kujawa2016effects; @byun2021extensive]. Changes to the underlying database and a rebuilding of the API required new data objects within the R package to more closely align to the Neotoma data model [@grimm2008neotoma]. The broad user community for Neotoma [@williams2018neotoma; @goring2018nexus] requires a toolset that can access and manage data for each of the more than 40 dataset types within Neotoma and so extensive metadata must be accessed for each record. This package conforms to a `tidyverse` [@wickham2019tidyverse] approach for data management, with data objects that more closely resemble the underlying data model within Neotoma (https://open.neotomadb.org/db_schema). Most importantly the `neotoma2` package provides a toolset for paleoecologists, ecologists, conservation ecologists, archaeologists, and others, to access and examine the broad range of fossil data contained within the Neotoma Paleoecology Database.
 
-Data from the Neotoma Database can be accessed through the public API (https://api.neotomadb.org) or a Postgres database snapshot with a database client. The R package will simplify many of the operations required to assemble and manipulate datasets for statistical analysis, and provides a number of functions that will support researchers as part of their analytic workflows.
+Data from the Neotoma Database can be accessed through the public API (https://api.neotomadb.org) or a Postgres database snapshot with a database client [@williams2018neotoma] or through the EarthLife Consortium API [@uhen2021earthlife]. The R package will simplify many of the operations required to assemble and manipulate datasets, and provides functions thatsupport researchers as part of their analytic workflows by linking Neotoma directly to packages within R used for ecological and earth science research, data visualization and statistical analysis.
 
 # Acknowledgements
 
