### 2023-08-07 NeotomaDB/neotoma2#32 @SimonGoring
Added additional context about the utility and need for the package that references other avenues for obtaining data from Neotoma.
