### 2023-08-09 1ecea8a
Updated the README to build vignettes on installation.
In response to issue #29. A small change.
