### 2023-08-09 b28f56a
Adding pkgdown build of package vignette.
In response to issue #30, have now built the pkgdown vignette and added it to the repository.
