### 2023-08-09 NeotomaDB/neotoma2#30 @SimonGoring
Thank you for this suggestion. I have added the change (in develop currently). Will be merging shortly.
