### 2023-08-09 7e312c6
Added email contact for development
There is now email contact information in the `DESCRIPTION` file and in the `README`. This addresses issue #31. Thanks!
