### 2023-08-09 NeotomaDB/neotoma2#31 @SimonGoring
Emails added to devel branch. CLosing this issue.
