### 2023-08-09 11d4a0c
Added a sentence to the paper and README to explain the end of `neotoma`
In reference to openjournals/joss-reviews issues #5561 [link](https://github.com/openjournals/joss-reviews/issues/5561) we have added a sentence to the README and to the paper.md file respectively to explain why the new package was required.
