### 2023-08-09 NeotomaDB/neotoma2#33 @SimonGoring
With respect to parallelization, this is a great suggestion, however, at present the R package depends on [a `while` loop](https://github.com/NeotomaDB/neotoma2/blob/production/R/parseURL.R#L144) to parse through the set of results (based on paging through offset/limit values).

Because of this, we send off smaller calls to `get_sites()` requesting 50 sites at a time until a result-set comes in empty (the case when the offset > the result set).

The Neotoma API does not, at present, provide a method for understanding the size of the result set, and so there's not a simple way to solve this problem across the set of queries passed to the API servers.

My sense is that there are ways around this (possibly using the [`future` package](https://future.futureverse.org/)?) but at present I don't think there's a simple solution.

That said, it might be easier to do this for the `build_sites()` methods (for example), and we'll look into it, but I'd like to put this option beyond the scope of the current review for JOSS.
