### 2023-08-09 72a4e09
Quality of life updates to specimens and parseURL
Just linting and adding some tests. No changes to any function operation.
