### 2023-08-14 bafd932
fixed some small bugs; fix that if nchar() longer than 2000, treat as 414 error
added get_stats() function
