+++ R/parseURL.R
@@ -12,14 +12,14 @@
 #' @description An internal helper function used to connect to the Neotoma API
 #' in a standard manner, and to provide basic validation of any response.
 #' @param x The HTTP/S path for the particular API call.
-#' @param use Uses the Neotoma server by default ("neotoma"), but supports
-#' either the development API server ("dev") or a local server ("local").
+#' @param use Uses the Neotoma server by default ("neotoma"), but supports either the
+#' development API server ("dev") or a local server ("local").
 #' @param all_data If TRUE return all possible API calls
 #' @param ... Any query parameters passed from the calling function.
 #' @returns `list` with cleaned and parsed data from HTTP request
 #' @export
 parseURL <- function(x, use = "neotoma", all_data = FALSE, ...) { # nolint
-
+  
   cleanNull <- function(x, fn = function(x) if (is.null(x)) NA else x) { # nolint
     if (is.list(x)) {
       lapply(x, cleanNull, fn)
@@ -27,46 +27,74 @@ parseURL <- function(x, use = "neotoma", all_data = FALSE, ...) { # nolint
       fn(x)
     }
   }
-
+  
   # Assign the API host location:
   if (!Sys.getenv("APIPOINT") == "") {
     use <- Sys.getenv("APIPOINT")
   }
-
+  
   baseurl <- switch(use,
                     "dev" = "https://api-dev.neotomadb.org/v2.0/",
                     "neotoma" = "https://api.neotomadb.org/v2.0/",
                     "local" = "http://localhost:3005/v2.0/",
                     use)
-
+  
   query <- list(...)
+
   if (all_data == FALSE) {
-    response <- httr::GET(paste0(baseurl, x),
-                          add_headers("User-Agent" = "neotoma2 R package"),
-                          query = query)
+    try(
+      response <- httr::GET(paste0(baseurl, x),
+                            add_headers("User-Agent" = "neotoma2 R package"),
+                            query = query)
+    )
+
+    if (inherits(response, "try-error")) {
+      # Handle the SSL error
+      error_message <- conditionMessage(response)
+      
+      if (grepl("SSL certificate", error_message, ignore.case = TRUE)) {
+        # Handle SSL certificate-related errors
+        stop("SSL certificate error:", error_message, "\n Please contact the Neotoma Team")
+      }
+    }
+    
     if (response$status_code == 414) {
       # The 414 error is a URL that is too long. This is a lazy way to manage
       # the choice between a POST and GET call.
       # Function with POST (Use this once server issue is resolved)
-      query <- list(...)
-      args <- x
-      new_url <- newURL(baseurl, args, ...)
-      body <- parsebody(args, ...)
-      response <- httr::POST(new_url,
+      new_url <- newURL(baseurl, x, ...)
+      body <- parsebody(x, all_data=FALSE, ...)
+      try(
+        response <- httr::POST(new_url,
                              body = body,
                              add_headers("User-Agent" = "neotoma2 R package"),
                              httr::content_type("application/json"))
-      warning("To get the complete data, use all_data = TRUE. 
-        Returned the first 25 elements.")
+      )
+      if (inherits(response, "try-error")) {
+        # Handle the SSL error
+        error_message <- conditionMessage(response)
+        
+        if (grepl("SSL certificate", error_message, ignore.case = TRUE)) {
+          # Handle SSL certificate-related errors
+          stop("SSL certificate error:", error_message, "\n Please contact the Neotoma Team")
+        }
+      }
+      # Break if we can't connect:
+      stop_for_status(response,
+                      task = "Could not connect to the Neotoma API.
+                    Check that the path is valid, and check the current
+                     status of the Neotoma API services at
+                      http://data.neotomadb.org")
+      warning("To get the complete data, use all_data = TRUE.")
     }
-
+    
     # Break if we can't connect:
[…]
