+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: neotoma2
 Title: Working with the Neotoma Paleoecology Database
-Date: 2023-04-17
-Version: 1.0.0
+Date: 2023-08-14
+Version: 1.0.2
 Authors@R:
     c(person(given = "Dominguez Vidana",
              family = "Socorro",
@@ -24,6 +24,7 @@ RoxygenNote: 7.2.3
 Suggests:
     testthat,
     knitr,
+    pkgdown,
     rmarkdown,
     Bchron,
     ggplot2,
@@ -48,4 +49,4 @@ Imports:
     wk,
     uuid,
     tidyr
-VignetteBuilder: knitr
+VignetteBuilder: pkgdown
