### 2023-08-14 258a3f3
get stats function
