+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: neotoma2
 Title: Working with the Neotoma Paleoecology Database
-Date: 2023-08-14
-Version: 1.0.2
+Date: 2023-08-16
+Version: 1.0.3
 Authors@R:
     c(person(given = "Dominguez Vidana",
              family = "Socorro",
@@ -49,4 +49,4 @@ Imports:
     wk,
     uuid,
     tidyr
-VignetteBuilder: pkgdown
+VignetteBuilder: knitr
