### 2023-08-17 94e4830
Updated News and cran-comments.md
