+++ NEWS.md
@@ -1,3 +1,8 @@
 # neotoma2 1.0.0
 
 * Added a `NEWS.md` file to track changes to the package.
+
+
+# neotoma2 1.0.3
+
+Added JOSS Article PDF and vignette under inst/doc
