### 2023-08-22 NeotomaDB/neotoma2#34 @sedv8808
get_stats() function has been created
