### 2023-08-22 #5561 @sedv8808
Hello, I've pushed the newest version 1.0.3 of `neotoma2` which addresses the comments of the review and other bugs that we have fixed.

This is ready for review @crvernon

Thank you,

Socorro
