### 2023-08-23 NeotomaDB/neotoma2#34 @kanishkan91
@sedv8808 Could you add a line in the documentation for this new function? Just what it does and how it can be used
