### 2023-08-23 #5561 @kanishkan91
@crvernon All looks good to me. I have requested that the authors (@sedv8808 and @SimonGoring) add a few lines to the documentation for the new get_stats function. Otherwise its good to go from my side
