### 2023-08-24 119cab4
Updated the CITATION and improved documentation for get_stats()
Changed the bib entry for the citation to `bibentry`.
