### 2023-08-28 NeotomaDB/neotoma2#36 @mhesselbarth
JOSS review (optional): Add link to rendered vignettes to docs
It would be nice if the link in the `README` provides a rendered HTML vignette directly. Maybe using Articles ([more information here](https://r-pkgs.org/vignettes.html)) would be nice. However, not necessarily required, just a suggestion.

https://github.com/openjournals/joss-reviews/issues/5561
