### 2023-08-28 NeotomaDB/neotoma2#37 @mhesselbarth
JOSS review: Manuscript
Lines 13-17: As far as I understand the manuscript and the vignettes, the `set_sites()` function can be used to create sites locally based on data not available in the database yet. However, currently in the manuscript it is unclear if the function also allows to upload these sites to the database ("create new records"). Please rephrase and specify.

Lines 30-33: Please specify a bit further how the package " [...] conforms to a `tidyverse` approach [...]" and how this relates to the Neotoma data model.

https://github.com/openjournals/joss-reviews/issues/5561
