### 2023-08-28 NeotomaDB/neotoma2#38 @mhesselbarth
JOSS review (optional): Code coverage
The automated tests are very well written, however, the overall coverage could be higher (currently about 60%) and many functions are currently not tested.

https://github.com/openjournals/joss-reviews/issues/5561
