### 2023-08-28 NeotomaDB/neotoma2#39 @mhesselbarth
JOSS Review: Code of conduct
Refer to code of conduct guidelines in documentation (Suggestions: Somewhere in the *Tips for Contributing* section)

https://github.com/openjournals/joss-reviews/issues/5561
