### 2023-08-28 #5561 @mhesselbarth
@crvernon

I am mostly done with my review. I opened a few issues, all are either optional or minor. Overall, the software seems to be really useful and well programmed.

The documentation and paper are also very well written and introduce the software very approachable (for some minor issue, see #37).
