### 2023-08-28 NeotomaDB/neotoma2#40 @mhesselbarth
JOSS Review (optional): Make functions internal
I think there are a few functions that could be made internal `testNull()` (besides others).  In my personal experience, limiting the exported functions to the core functionality is often beneficial to no confuse less experienced users.

https://github.com/openjournals/joss-reviews/issues/5561
