### 2023-08-29 NeotomaDB/neotoma2#39 @SimonGoring
Hi @mhesselbarth, we can certainly add a link to (or reference) the JOSS Code of Conduct. There is already a [code of conduct] for the repo, that both covers the JOSS one (I believe) and goes a bit further. Is the point you're making to ensure that the Code is referenced within the vignettes/documentation itself? Or just in the README?

Thanks!
