### 2023-08-29 #5561 @SimonGoring
> @crvernon
>
> I am mostly done with my review. I opened a few issues, all are either optional or minor. Overall, the software seems to be really useful and well programmed.
>
> The documentation and paper are also very well written and introduce the software very approachable (for some minor issue, see #37).

I believe this is a reference to the NeotomaDB/neotoma2 Issue #37.
