### 2023-08-29 7905670
Minor text additions to improve article clarity.
In response to @mhesselbarth's comments in Issue #37 I have edited the Summary and body of the article.

To address the question about the `set_*()` functions I have indicated that the intention is to support upload, but that that the functionality (on the Neotoma Database side) is not present yet:

> The `set_*()` functions are intended as a precursor to utilities to upload data directly to Neotoma, although this functionality is not yet available.

The question about how the package conforms to a `tidyverse` model is explained here:

> This pa
[…]
