+++ paper.bib
@@ -52,6 +52,14 @@ volume={13},
 number={2},
 year={2021}
 }
+
+@Manual{oksanen2022vegan,
+title = {vegan: Community Ecology Package},
+author = {Jari Oksanen and Gavin L. Simpson and F. Guillaume Blanchet and Roeland Kindt and Pierre Legendre and Peter R. Minchin and R.B. O'Hara and Peter Solymos and M. Henry H. Stevens and Eduard Szoecs and Helene Wagner and Matt Barbour and Michael Bedward and Ben Bolker and Daniel Borcard and Gustavo Carvalho and Michael Chirico and Miquel {De Caceres} and Sebastien Durand and Heloisa Beatriz Antoniazi Evangelista and Rich FitzJohn and Michael Friendly and Brendan Furneaux and Geoffrey Hannigan and Mark O. Hill and Leo Lahti and Dan McGlinn and Marie-Helene Ouellette and Eduardo {Ribeiro Cunha} and Tyler Smith and Adrian Stier and Cajo J.F. {Ter Braak} and James Weedon},
+year = {2022},
+note = {R package version 2.6-2},
+url = {https://CRAN.R-project.org/package=vegan},
+}
   
 @article{williams2018neotoma,
   title={The Neotoma Paleoecology Database, a multiproxy, international, community-curated data resource},
