### 2023-08-29 6771e8a
Change functions to be internal
