+++ R/use_na.R
@@ -5,7 +5,8 @@
 #' @param type A character string with values either \code{char} or \code{int}.
 #' @importFrom purrr map
 #' @returns object converted to `NA_character` or `NA_integer`
-#' @export
+#' @keywords internal
+#' @noRd
 use_na <- function(x, type) {
   if (is.na(x)) {
     return(switch(type,
