+++ R/fix_nulls.r
@@ -3,7 +3,8 @@
 #' as NULL values. This function turns the NULL values into logical NAs.
 #' @param x item that has null objects.
 #' @returns `list` with `NA`s instead of `null` objects
-#' @export
+#' @keywords internal
+#' @noRd
 fix_null <- function(x) {
     for (i in seq_len(length(x))) {
       if (is.null(x[[i]])) {
