+++ man/cleanNULL.Rd
@@ -1,18 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/cleanNULL.R
-\name{cleanNULL}
-\alias{cleanNULL}
-\title{Clean NULL values}
-\usage{
-cleanNULL(x)
-}
-\arguments{
-\item{x}{An element that may or may not have NULL values.}
-}
-\value{
-parsed \code{list} where NULL values are changed to NA
-}
-\description{
-Pass an object and convert all
-\code{NULL} elements to \code{NA}.
-}
