+++ NAMESPACE
@@ -1,9 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
 S3method(check_contacts,contacts)
-S3method(clean,collunits)
-S3method(clean,datasets)
-S3method(clean,sites)
 S3method(filter,sites)
 S3method(get_contacts,default)
 S3method(get_contacts,numeric)
@@ -39,14 +36,11 @@ export(check_contacts)
 export(chroncontrols)
 export(chronologies)
 export(cite_data)
-export(clean)
-export(cleanNULL)
 export(collunits)
 export(coordinates)
 export(datasets)
 export(doi)
 export(filter)
-export(fix_null)
 export(get_contacts)
 export(get_datasets)
 export(get_downloads)
@@ -74,10 +68,8 @@ export(set_site)
 export(showMatch)
 export(specimens)
 export(taxa)
-export(testNull)
 export(toJSON)
 export(toWide)
-export(use_na)
 exportClasses(author)
 exportClasses(authors)
 exportClasses(chronologies)
