### 2023-08-29 77a7400
Changed set_ functions documentation to showcase they are not for uploading data into the DB as per JOSS request
