### 2023-08-29 bce8a57
internal function
