### 2023-08-29 23226a9
remove test of example as it takes longer than the allowed time
