### 2023-08-29 bddb6fa
remove clean() from example as it is an internal function
