### 2023-08-29 NeotomaDB/neotoma2#40 @sedv8808
Commit 6771e8a addresses this issue
