### 2023-08-29 NeotomaDB/neotoma2#36 @sedv8808
Commit 199e3d5 addresses this issue
