### 2023-08-30 NeotomaDB/neotoma2#39 @mhesselbarth
Hey,

Yes, its great you already have a Code of Conduct document. I think just adding one sentence in the "Tips for Contributing" section would be sufficient. Something along the lines of "If you want to contribute to this project, please be aware of our [Code of Conduct](https://github.com/NeotomaDB/neotoma2/blob/production/CODE_OF_CONDUCT.md)".
