### 2023-10-16 #5561 @mhesselbarth
I just went ahead and closed all of my issues that were actually addressed. There are still 1,2 open - but they are more suggestions anyhow.

So ready to make another pass from my side.
