### 2023-11-06 #5561 @crvernon
:wave: @SimonGoring

Great work on this submission!

Here are my comments related to your paper:

- [ ] LINES 14-15:  I believe the use of the wildcard asterisk makes reading a bit fuzzy.  You can simply state "...create new records using get and set functions (e.g., `get_sites()`, `set_sites()`)."
- [ ] Throughout, related to the above comment.  Just stating "set or get functions" instead of using the asterisks should still be good.
- [ ] Throughout, for package names and functions, use the backtick formatting to distinguish them from the standard text.
- [ ] LINE 47:  "thatsupport" should be two words
- [ ] See to missing DOI identified in check above:  ```10.21425/f5fbg50711 may be a valid DOI for title: The EarthLife Consortium API: an extensible, open-source service for accessing fossil data and taxonomies from multiple community paleodata resources```

After you address these items, I believe we can move forward with the next steps.  Let me know if you have any questions!
