### 2023-11-06 6bf9249
Auto stash before rebase of "origin/production"
Added DOI for ELC and updated changes referenced in https://github.com/openjournals/joss-reviews/issues/5561#issuecomment-1794954673
