+++ paper.md
@@ -27,7 +27,7 @@ bibliography: paper.bib
 
 # Summary
 
-The `neotoma2` R package is a tool to access and manipulate data from the Neotoma Paleoecology Database (https://www.neotomadb.org) within the R environment. Neotoma is a community curated paleoecological data resource [@williams2018neotoma], containing nearly 9 million unique observations of paleoecological proxies with global coverage from 37 constituent databases. The package uses the Neotoma API v2.0 [@goring2023api] as a tool to import records from the Neotoma database, allowing researchers to examine taxonomic, spatial and temporal patterns across space and time over the last 5.4 million years. The R package allows researchers to download, and create new records using `get_*()` (e.g., `get_sites()`) and `set_*()` functions (e.g., `set_sites()`) respectively. This provides researchers with the opportunity to develop dynamic workflows that include data generated locally, and not yet submitted to the Neotoma database. The `set_*()` functions are intended as a precursor to utilities to upload data directly to Neotoma, although this functionality is not yet available.
+The `neotoma2` R package is a tool to access and manipulate data from the Neotoma Paleoecology Database (https://www.neotomadb.org) within the R environment. Neotoma is a community curated paleoecological data resource [@williams2018neotoma], containing nearly 9 million unique observations of paleoecological proxies with global coverage from 37 constituent databases. The package uses the Neotoma API v2.0 [@goring2023api] as a tool to import records from the Neotoma database, allowing researchers to examine taxonomic, spatial and temporal patterns across space and time over the last 5.4 million years. The R package allows researchers to download, and create new records using `get` and `set` functions (e.g., `get_sites()`, `set_sites()`) respectively. This provides researchers with the opportunity to develop dynamic workflows that include data generated locally, and not yet submitted to the Neotoma database. The `set` functions are intended as a precursor to utilities to upload data directly to Neotoma, although this functionality is not yet available.
 
 The `neotoma2` R package has been under dynamic development for over a year, but has been used for teaching and training [@Goring2023APD]. This release of the `neotoma2` R package is a clean release of the package, with all of the core features provided and extensive test coverage implemented.
 
@@ -37,7 +37,7 @@ The `neotoma` R package [@goring2015neotoma] leveraged the Neotoma Paleoeocology
 
 The broad user community for Neotoma [@williams2018neotoma; @goring2018nexus] requires a toolset that can access and manage data for each of the more than 40 dataset types within Neotoma and so extensive metadata must be accessed for each record. This package conforms to a `tidyverse` [@wickham2019tidyverse] approach for data management, with methods and data objects that are suited to piping using the `%>%` (or `|>`) pipe convention and with implementations for filtering and other common `dplyr` methods. The package also now uses "long" `data.frames` and `tibbles` by default, using the `toWide()` function to transform data into "wide" tables for use with common ecological data packages such as `vegan` [@oksanen2022vegan].  Data objects in the `neotoma2` package now more closely resemble the underlying data model within Neotoma (https://open.neotomadb.org/db_schema) than in the previous package. Most importantly the `neotoma2` package provides a toolset for paleoecologists, ecologists, conservation ecologists, archaeologists, and others, to access and examine the broad range of fossil data contained within the Neotoma Paleoecology Database.
 
-Data from the Neotoma Database can be accessed through the public API (https://api.neotomadb.org) or a Postgres database snapshot with a database client [@williams2018neotoma] or through the EarthLife Consortium API [@uhen2021earthlife]. The R package will simplify many of the operations required to assemble and manipulate datasets, and provides functions thatsupport researchers as part of their analytic workflows by linking Neotoma directly to packages within R used for ecological and earth science research, data visualization and statistical analysis.
[…]
