### 2023-11-06 #5561 @SimonGoring
I've addressed all the changed described above in the neotoma2 repo commit 6bf9249feea0a741043352dbf031a450563e9825, and then followed it up with a change of the date on the paper.
