### 2023-11-16 4c49240
Create CITATION.cff
