+++ CITATION.cff
@@ -1,0 +1,14 @@
+cff-version: 1.2.0
+message: "If you use this software, please cite it as below."
+authors:
+- family-names: "Dominguez"
+  given-names: "Socorro"
+  orcid: "https://orcid.org/0000-0002-7926-4935"
+- family-names: "Goring"
+  given-names: "Simon"
+  orcid: "https://orcid.org/0000-0002-2700-4605"
+title: "The neotoma2 R package for the Neotoma Paleoecology Database"
+version: 1.0.3
+doi: 10.5281/zenodo.10144558
+date-released: 2023-11-16
+url: "https://github.com/NeotomaDB/neotoma2"
