+++ CITATION.cff
@@ -9,6 +9,6 @@ authors:
   orcid: "https://orcid.org/0000-0002-2700-4605"
 title: "neotoma2: An R package to access data from the Neotoma Paleoecology Database"
 version: 1.0.3
-doi: 10.5281/zenodo.10144558
+doi: 10.5281/zenodo.7892992
 date-released: 2023-11-16
 url: "https://github.com/NeotomaDB/neotoma2"
