### 2023-11-16 #5561 @SimonGoring
Hi, got everything done now. Added the CITATION.cff after the release, because there's already a citation bib file in the `inst` folder. I can re-make the release if you'd like.

Otherwise, the current DOI is https://doi.org/10.5281/zenodo.10144558
