### 2023-11-21 #5561 @crvernon
Hi @SimonGoring , you can just edit the title and authors of your Zenodo record (https://doi.org/10.5281/zenodo.10144558) to match the title and authors for this paper.  That should be sufficient.  Just let me know when it is done and we can move forward.
