### 2023-11-22 #5561 @SimonGoring
I've updated the release with a proper CITATION.cff file as well. DOI: https://doi.org/10.5281/zenodo.10198238
