### 2023-11-27 82d0770
Spelling and formatting for JOSS
I believe these are the final revisions neccessary for https://github.com/openjournals/joss-reviews/issues/5561

Minor spelling correction and proper bib formatting to ensure accurate capitalizaiton.
