+++ paper.bib
@@ -1,5 +1,5 @@
 @article{goring2015neotoma,
-  title={{neotoma}: A programmatic interface to the Neotoma Paleoecological Database},
+  title={{neotoma}: A programmatic interface to the {N}eotoma {P}aleoecological {D}atabase},
   author={Goring, Simon and Dawson, Andria and Simpson, Gavin and Ram, Karthik and Graham, Russ and Grimm, Eric and Williams, John W.},
   journal={Open Quaternary},
   volume={1},
@@ -10,7 +10,7 @@
 }
 
 @article{byun2021extensive,
-title={Extensive wetland development in mid-latitude North America during the B{\o}lling--Aller{\o}d},
+title={Extensive wetland development in mid-latitude {N}orth {A}merica during the {B}{\o}lling--{A}ller{\o}d},
 author={Byun, Eunji and Sato, Hiromitsu and Cowling, Sharon A and Finkelstein, Sarah A},
 journal={Nature Geoscience},
 volume={14},
@@ -22,7 +22,7 @@ doi={10.1038/s41561-020-00670-4}
 }
 
 @article{kujawa2016effects,
-title={The effects of anthropogenic land cover change on pollen-vegetation relationships in the American Midwest},
+title={The effects of anthropogenic land cover change on pollen-vegetation relationships in the {A}merican {M}idwest},
 author={Kujawa, Ellen Ruth and Goring, Simon and Dawson, Andria and Calcote, Randy and Grimm, Eric C and Hotchkiss, Sara C and Jackson, Stephen T and Lynch, Elizabeth A and McLachlan, Jason and St-Jacques, Jeannine-Marie and {Umbanhowar Jr.}, Charles and Williams, John W.},
 journal={Anthropocene},
 volume={15},
@@ -45,7 +45,7 @@ doi={10.1073/pnas.2201946119}
 }
 
 @article{uhen2021earthlife,
-title={The EarthLife Consortium API: an extensible, open-source service for accessing fossil data and taxonomies from multiple community paleodata resources},
+title={The {E}arth{L}ife {C}onsortium {API}: an extensible, open-source service for accessing fossil data and taxonomies from multiple community paleodata resources},
 author={Uhen, Mark D and Buckland, Philip I and Goring, Simon J and Jenkins, Julian P and Williams, John W},
 journal={Frontiers of Biogeography},
 volume={13},
@@ -63,7 +63,7 @@ url = {https://CRAN.R-project.org/package=vegan},
 }
   
 @article{williams2018neotoma,
-  title={The Neotoma Paleoecology Database, a multiproxy, international, community-curated data resource},
+  title={The {N}eotoma {P}aleoecology {D}atabase, a multiproxy, international, community-curated data resource},
   author={Williams, John W and Grimm, Eric C and Blois, Jessica L and Charles, Donald F and 
   Davis, Edward B and Goring, Simon J and Graham, Russell W and Smith, Alison J and
   Anderson, Michael and Arroyo-Cabrales, Joaquin and Ashworth, Allan C. and Betancourt, Julio L. and
@@ -86,7 +86,7 @@ url = {https://CRAN.R-project.org/package=vegan},
   @book{goring2018nexus,
     place={Cambridge},
     series={Elements of Paleontology},
-    title={The Neotoma Paleoecology Database: A Research Outreach Nexus},
+    title={The {N}eotoma {P}aleoecology {D}atabase: A Research Outreach Nexus},
     doi={10.1017/9781108681582},
     publisher={Cambridge University Press},
[…]
