+++ paper.md
@@ -33,9 +33,9 @@ The `neotoma2` R package has been under dynamic development for over a year, but
 
 # Statement of Need
 
-The `neotoma` R package [@goring2015neotoma] leveraged the Neotoma Paleoeocology Database v1.0 API and had been one of the primary tools for researchers working with data from Neotoma [@wang2023plants;@kujawa2016effects; @byun2021extensive]. Changes to the underlying database and a rebuilding of the API required new data objects within the R package to more closely align to the Neotoma data model [@grimm2008neotoma]. Additionally, the original v1.0 API that was accessed by the `neotoma` pacakge was deprecated in 2020, meaning the `neotoma` package could no longer access data from Neotoma.
+The `neotoma` R package [@goring2015neotoma] leveraged the Neotoma Paleoeocology Database v1.0 API and had been one of the primary tools for researchers working with data from Neotoma [@wang2023plants;@kujawa2016effects; @byun2021extensive]. Changes to the underlying database and a rebuilding of the API required new data objects within the R package to more closely align to the Neotoma data model [@grimm2008neotoma]. Additionally, the original v1.0 API that was accessed by the `neotoma` package was deprecated in 2020, meaning the `neotoma` package could no longer access data from Neotoma.
 
-The broad user community for Neotoma [@williams2018neotoma; @goring2018nexus] requires a toolset that can access and manage data for each of the more than 40 dataset types within Neotoma and so extensive metadata must be accessed for each record. This package conforms to a `tidyverse` [@wickham2019tidyverse] approach for data management, with methods and data objects that are suited to piping using the `%>%` (or `|>`) pipe convention and with implementations for filtering and other common `dplyr` methods. The package also now uses "long" `data.frames` and `tibbles` by default, using the `toWide()` function to transform data into "wide" tables for use with common ecological data packages such as `vegan` [@oksanen2022vegan].  Data objects in the `neotoma2` package now more closely resemble the underlying data model within Neotoma (https://open.neotomadb.org/db_schema) than in the previous package. Most importantly the `neotoma2` package provides a toolset for paleoecologists, ecologists, conservation ecologists, archaeologists, and others, to access and examine the broad range of fossil data contained within the Neotoma Paleoecology Database.
+The broad user community for Neotoma [@williams2018neotoma; @goring2018nexus] requires a toolset that can access and manage data for each of the more than 40 dataset types within Neotoma and so extensive metadata must be accessed for each record. The `neotoma2` package conforms to a `tidyverse` [@wickham2019tidyverse] approach for data management, with methods and data objects that are suited to piping using the `%>%` (or `|>`) pipe convention and with implementations for filtering and other common `dplyr` methods. The package also now uses "long" `data.frames` and `tibbles` by default, using the `toWide()` function to transform data into "wide" tables for use with common ecological data packages such as `vegan` [@oksanen2022vegan].  Data objects in the `neotoma2` package now more closely resemble the underlying data model within Neotoma (https://open.neotomadb.org/dbschema) than in the previous package. Most importantly the `neotoma2` package provides a toolset for paleoecologists, ecologists, conservation ecologists, archaeologists, and others, to access and examine the broad range of fossil data contained within the Neotoma Paleoecology Database.
 
 Data from the Neotoma Database can be accessed through the public API (https://api.neotomadb.org) or a PostgreSQL database snapshot with a database client [@williams2018neotoma; https://www.neotomadb.org/data/db-snapshots] or through the EarthLife Consortium API [@uhen2021earthlife]. The R package will simplify many of the operations required to assemble and manipulate datasets, and provides functions that support researchers as part of their analytic workflows by linking Neotoma directly to packages within R used for ecological and earth science research, data visualization and statistical analysis.
 
